refactor: split HypergraphQuery trait into focused submodules
