+++ src/core/query/mod.rs
@@ -1,6 +1,13 @@
 mod hypergraph_impl;
+mod lookups;
+mod paths;
+mod projections;
+mod properties;
+mod structural;
+mod trait_def;
+mod traversal;
+
 #[cfg(feature = "persistence")]
 mod persistent_impl;
-mod trait_def;
 
 pub use trait_def::HypergraphQuery;
